Extract unified JdbcMetering from JdbcAvroMetering and JdbcParquetMetering
The two metering classes were 100% identical except for the log prefix
("jdbcavroio" vs "jdbcparquetio"). Introduce JdbcMetering parameterized
by a format name string and update all callers.

JdbcAvroMetering and JdbcParquetMetering are retained as deprecated
thin subclasses for backwards compatibility.
